Updated Code

+++ backend/src/main/java/com/petstore/repository/PetRepository.java
@@ -1,7 +1,8 @@
 package com.petstore.repository;
 
-import com.petstore.model.Pet;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PetRepository extends JpaRepository<Pet, Long> {
+import com.petstore.model.Pet;
+
+public interface PetRepository extends JpaRepository<Pet, Integer> {
 }
